src/requirements.txt
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.10.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.10
django-admin-sortable==2.2.4
django-appconf==1.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.1
django-braces==1.15.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==1.1.0
django-intl-tel-input==0.3.1
django-intl-tel-input2==0.2.0
django-jquery==3.1.0
django-location-field==2.1.0
django-mathfilters==1.0.0
django-money==2.1
django-moneyfield==0.2.1
django-mssql-backend==2.8.1
django-multiselectfield==0.1.12
django-select2==7.9.0
django-settings-export==1.2.1
django-tables2==2.4.1
django-utils-six==2.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-filters==0.11.1
easy-thumbnails==2.8
et-xmlfile==1.1.0
importlib-metadata==4.8.2
jdcal==1.4.1
lxml==4.7.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
pip-review==1.1.0
psycopg2==2.9.2
py-moneyed==1.2
pyodbc==4.0.32
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
reportlab==3.6.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.1.0
tablib==3.1.0
tinycss2==1.1.1
typing_extensions==4.0.0
unicodecsv==0.14.1
webencodings==0.5.1
Werkzeug==2.0.2
wfastcgi==3.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
xmlrunner==1.7.7
zipp==3.6.0
